@@ -4,92 +4,96 @@
   "description": "Web app for Tab for a Cause.",
   "private": true,
   "devDependencies": {
-    "babel-plugin-module-resolver": "^2.7.0",
-    "babel-plugin-relay": "dev",
+    "babel-plugin-module-resolver": "^3.0.0-beta.0",
+    "babel-plugin-relay": "^1.0.1-rc.4",
     "codecov": "^2.1.0",
-    "dotenv-extended": "^1.0.4",
-    "enzyme": "^2.8.0",
+    "dotenv-extended": "^2.0.0",
+    "enzyme": "^2.8.2",
     "if-env": "^1.0.0",
     "lodash": "^4.17.4",
-    "react-addons-test-utils": "^15.4.2",
-    "relay-compiler": "1.0.0-rc.2"
+    "npm-run-all": "^4.0.2",
+    "react-test-renderer": "^15.5.4",
+    "relay-compiler": "^1.0.0-rc.4"
   },
   "dependencies": {
     "amazon-cognito-identity-js": "^1.16.0",
-    "autoprefixer": "6.5.1",
+    "autoprefixer": "^7.1.0",
     "babel-core": "^6.23.1",
-    "babel-eslint": "7.1.1",
-    "babel-jest": "17.0.2",
-    "babel-loader": "^6.3.2",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^20.0.1",
+    "babel-loader": "^7.0.0",
     "babel-plugin-react-intl": "^2.3.1",
     "babel-plugin-react-transform": "^2.0.2",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-0": "^6.22.0",
-    "babel-relay-plugin": "^0.10.0",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
+    "babel-relay-plugin": "^0.11.0",
+    "case-sensitive-paths-webpack-plugin": "^2.0.0",
     "chalk": "1.1.3",
     "classnames": "^2.2.5",
     "connect-history-api-fallback": "1.3.0",
-    "cross-spawn": "4.0.2",
-    "css-loader": "0.26.0",
-    "detect-port": "1.0.1",
-    "eslint": "3.8.1",
-    "eslint-config-react-app": "^0.5.0",
-    "eslint-loader": "1.6.0",
-    "eslint-plugin-flowtype": "2.21.0",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.4.1",
-    "extract-text-webpack-plugin": "1.0.1",
+    "cross-spawn": "^5.1.0",
+    "css-loader": "^0.28.1",
+    "detect-port": "^1.1.2",
+    "eslint": "^3.19.0",
+    "eslint-config-react-app": "^0.6.2",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-flowtype": "^2.33.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-react": "^7.0.1",
+    "extract-text-webpack-plugin": "^2.1.0",
     "favicons-webpack-plugin": "0.0.7",
-    "file-loader": "0.9.0",
-    "filesize": "3.3.0",
+    "file-loader": "^0.11.1",
+    "filesize": "^3.5.9",
     "font-awesome": "^4.7.0",
-    "fs-extra": "0.30.0",
+    "fs-extra": "^3.0.1",
     "gzip-size": "3.0.0",
-    "html-webpack-plugin": "2.24.0",
-    "http-proxy-middleware": "0.17.2",
-    "jest": "17.0.2",
+    "html-webpack-plugin": "^2.28.0",
+    "http-proxy-middleware": "^0.17.4",
+    "jest": "^20.0.1",
     "json-loader": "0.5.4",
-    "material-ui": "^0.17.1",
+    "material-ui": "^0.18.1",
     "moment": "^2.18.1",
     "nodemon": "^1.11.0",
-    "normalize.css": "^5.0.0",
-    "object-assign": "4.1.0",
-    "path-exists": "2.1.0",
-    "postcss-loader": "1.0.0",
+    "normalize.css": "^7.0.0",
+    "object-assign": "^4.1.1",
+    "path-exists": "^3.0.0",
+    "postcss-loader": "^2.0.5",
     "precss": "^1.4.0",
     "promise": "7.1.1",
+    "prop-types": "^15.5.10",
     "react": "^15.4.2",
     "react-color": "^2.11.7",
-    "react-dev-utils": "^0.4.2",
+    "react-dev-utils": "^0.5.2",
     "react-dom": "^15.4.2",
     "react-dropdown": "^1.2.0",
     "react-hot-loader": "3.0.0-beta.6",
     "react-intl": "^2.2.3",
-    "react-relay": "dev",
+    "react-relay": "^1.0.0-rc.4",
     "react-router": "^3.0.0",
     "react-tap-event-plugin": "^2.0.1",
     "react-transform-catch-errors": "^1.0.2",
     "react-transform-hmr": "^1.0.4",
-    "recursive-readdir": "2.1.0",
+    "recursive-readdir": "^2.2.1",
     "redbox-react": "^1.3.3",
     "strip-ansi": "3.0.1",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
+    "style-loader": "^0.17.0",
+    "url-loader": "^0.5.8",
     "webpack": "^2.2.1",
     "webpack-dev-server": "^2.4.1",
     "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "2.0.2"
+    "whatwg-fetch": "^2.0.3"
   },
   "scripts": {
     "start": "yarn run relay && nodemon --watch data/schema.graphql scripts/start.js & yarn run relay:watch",
     "build": "node scripts/build.js",
     "relay": "relay-compiler --src ./src/ --schema ./data/schema.graphql",
     "relay:watch": "nodemon --watch src/ --watch data/schema.graphql --ignore **/__generated__/**/* --exec \"yarn run relay\"",
-    "test": "jest --env=jsdom --coverage && if-env NODE_ENV=test && codecov || echo 'Skipping test coverage logging.'",
+    "test": "npm-run-all -s test:run test:codecov",
+    "test:run": "jest --env=jsdom --coverage",
+    "test:codecov": "if-env NODE_ENV=test && codecov || echo 'Skipping test coverage logging.'",
     "test:watch": "jest --env=jsdom --watch",
     "build:langs": "node scripts/translate.js"
   },
