@@ -1,6 +1,6 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { goTo, goToDashboard } from 'navigation/navigation';
-
 import AppBar from 'material-ui/AppBar';
 import FontIcon from 'material-ui/FontIcon';
 import Menu from 'material-ui/Menu';
@@ -129,7 +129,7 @@ class Settings extends React.Component {
 }
 
 Settings.propTypes = {
-  user: React.PropTypes.object.isRequired
+  user: PropTypes.object.isRequired
 };
 
 export default Settings;
